% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binders.R
\name{bind_gram_cat}
\alias{bind_gram_cat}
\alias{bind_gender}
\alias{bind_infover}
\title{Binders functions}
\usage{
bind_gram_cat(df, col)

bind_gender(df, col)

bind_infover(df, col)
}
\arguments{
\item{df}{the dataframe}

\item{col}{the column with words}
}
\value{
a tibble
}
\description{
\describe{
  \item{bind_gram_cat}{bind the grammatical category of a word}
  \item{bind_gender}{bind the gender of a word (if appliable)}
  \item{bind_infover}{bind the verbal info of a verb (if appliable)}
}
}
\examples{
ad <- proustr::albertinedisparue
b <- tidytext::unnest_tokens(ad, mots, text) \%>\%
    count(mots) \%>\%
    slice(1:10)
bind_homograph(b, mots)
bind_gram_cat(b, mots)
}
