% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{count_letters}
\alias{count_letters}
\alias{count_homogr}
\alias{count_homoph}
\alias{count_phons}
\alias{count_syll}
\alias{count_morph}
\title{Counters}
\usage{
count_letters(df, col)

count_homogr(df, col)

count_homoph(df, col)

count_phons(df, col)

count_syll(df, col)

count_morph(df, col)
}
\arguments{
\item{df}{the dataframe}

\item{col}{the column with words}
}
\value{
a data.frame with words and a column with the specific count of element
}
\description{
\describe{
  \item{count_letters}{counts the number of letters in a word}
  \item{count_homogr}{counts the number of homograms of a word}
  \item{count_homoph}{counts the number of homographs of a word}
  \item{count_phons}{counts the number of phonems in a word}
  \item{count_syll}{counts the number of syllables in a word}
  \item{count_morph}{counts the number of morphems in a word}
}
}
