% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binders.R
\name{bind_homograph}
\alias{bind_homograph}
\alias{bind_homophon}
\alias{bind_phonem}
\alias{bind_morphem}
\alias{bind_gram_cat}
\title{Binders functions}
\usage{
bind_homograph(df, col)

bind_homophon(df, col)

bind_phonem(df, col)

bind_morphem(df, col)

bind_gram_cat(df, col)
}
\arguments{
\item{df}{the dataframe}

\item{col}{the column with words}
}
\value{
a tibble
}
\description{
Binders functions
}
\examples{
ad <- proustr::albertinedisparue
baka <- tidytext::unnest_tokens(ad, word, text) \%>\%
    count(word)
bind_homograph(baka, word)
bind_gram_cat(baka, word)
}
