% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{count_letters}
\alias{count_letters}
\alias{count_homogr}
\alias{count_homoph}
\alias{count_phons}
\alias{count_syll}
\alias{count_morph}
\title{Title}
\usage{
count_letters(df, col)

count_homogr(df, col)

count_homoph(df, col)

count_phons(df, col)

count_syll(df, col)

count_morph(df, col)
}
\arguments{
\item{col}{}
}
\description{
Title
}
