% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_lemme.R
\name{is_lemme}
\alias{is_lemme}
\title{Is the word a lemme?}
\usage{
is_lemme(df, col)
}
\arguments{
\item{df}{a data frame}

\item{col}{the column with words}
}
\value{
a tibble
}
\description{
Is the word a lemme?
}
\examples{

}
